Script to edit excel spreadsheets with openpyxl
This script extracts information from cells in a spreadsheet and writes them to empty columns. It also re-arranges the 
positions of various cells which makes the spreadsheet more suitable for downstream analysis. 

The spreadsheet contained raw pixel values of UV and non-UV images of nestling birds.

Would be great to hear people's suggestions on how script could be improved or made more succinct!